+++ Cargo.toml
@@ -35,6 +35,7 @@ members = [
   "resolvers/deno",
   "resolvers/node",
   "resolvers/npm_cache",
+  "resolvers/npm_installer",
   "runtime",
   "runtime/features",
   "runtime/permissions",
@@ -112,6 +113,7 @@ deno_bench_util = { version = "0.198.0", path = "./bench_util" }
 deno_features = { version = "0.1.0", path = "./runtime/features" }
 deno_lib = { version = "0.20.0", path = "./cli/lib" }
 deno_npm_cache = { version = "0.23.0", path = "./resolvers/npm_cache" }
+deno_npm_installer = { version = "0.0.1", path = "./resolvers/npm_installer" }
 deno_permissions = { version = "0.63.0", path = "./runtime/permissions" }
 deno_resolver = { version = "0.35.0", path = "./resolvers/deno" }
 deno_runtime = { version = "0.212.0", path = "./runtime" }
@@ -234,7 +236,7 @@ simd-json = "0.14.0"
 slab = "0.4"
 smallvec = "1.8"
 socket2 = { version = "0.5.3", features = ["all"] }
-sys_traits = "=0.1.9"
+sys_traits = "=0.1.10"
 tar = "=0.4.43"
 tempfile = "3.4.0"
 termcolor = "1.1.3"
@@ -516,3 +518,6 @@ opt-level = 3
 opt-level = 3
 [profile.release.package.deno_npm_cache]
 opt-level = 3
+
+[patch.crates-io]
+sys_traits = { path = "../sys_traits" }
