+++ cli/util/fs.rs
@@ -150,182 +150,6 @@ pub async fn remove_dir_all_if_exists(path: &Path) -> std::io::Result<()> {
   }
 }
 
-/// Clones a directory to another directory. The exact method
-/// is not guaranteed - it may be a hardlink, copy, or other platform-specific
-/// operation.
-///
-/// Note: Does not handle symlinks.
-pub fn clone_dir_recursive<
-  TSys: sys_traits::FsCopy
-    + sys_traits::FsCloneFile
-    + sys_traits::FsCloneFile
-    + sys_traits::FsCreateDir
-    + sys_traits::FsHardLink
-    + sys_traits::FsReadDir
-    + sys_traits::FsRemoveFile
-    + sys_traits::ThreadSleep,
->(
-  sys: &TSys,
-  from: &Path,
-  to: &Path,
-) -> Result<(), CopyDirRecursiveError> {
-  if cfg!(target_vendor = "apple") {
-    if let Some(parent) = to.parent() {
-      sys.fs_create_dir_all(parent)?;
-    }
-    // Try to clone the whole directory
-    if let Err(err) = sys.fs_clone_file(from, to) {
-      if !matches!(
-        err.kind(),
-        std::io::ErrorKind::AlreadyExists | std::io::ErrorKind::Unsupported
-      ) {
-        log::debug!(
-          "Failed to clone dir {:?} to {:?} via clonefile: {}",
-          from,
-          to,
-          err
-        );
-      }
-      // clonefile won't overwrite existing files, so if the dir exists
-      // we need to handle it recursively.
-      copy_dir_recursive(sys, from, to)?;
-    }
-  } else if let Err(e) = deno_npm_cache::hard_link_dir_recursive(sys, from, to)
-  {
-    log::debug!("Failed to hard link dir {:?} to {:?}: {}", from, to, e);
-    copy_dir_recursive(sys, from, to)?;
-  }
-
-  Ok(())
-}
-
-#[derive(Debug, thiserror::Error, deno_error::JsError)]
-pub enum CopyDirRecursiveError {
-  #[class(inherit)]
-  #[error("Creating {path}")]
-  Creating {
-    path: PathBuf,
-    #[source]
-    #[inherit]
-    source: Error,
-  },
-  #[class(inherit)]
-  #[error("Reading {path}")]
-  Reading {
-    path: PathBuf,
-    #[source]
-    #[inherit]
-    source: Error,
-  },
-  #[class(inherit)]
-  #[error("Dir {from} to {to}")]
-  Dir {
-    from: PathBuf,
-    to: PathBuf,
-    #[source]
-    #[inherit]
-    source: Box<Self>,
-  },
-  #[class(inherit)]
-  #[error("Copying {from} to {to}")]
-  Copying {
-    from: PathBuf,
-    to: PathBuf,
-    #[source]
-    #[inherit]
-    source: Error,
-  },
-  #[class(inherit)]
-  #[error(transparent)]
-  Other(#[from] Error),
-}
-
-/// Copies a directory to another directory.
-///
-/// Note: Does not handle symlinks.
-pub fn copy_dir_recursive<
-  TSys: sys_traits::FsCopy
-    + sys_traits::FsCloneFile
-    + sys_traits::FsCreateDir
-    + sys_traits::FsHardLink
-    + sys_traits::FsReadDir,
->(
-  sys: &TSys,
-  from: &Path,
-  to: &Path,
-) -> Result<(), CopyDirRecursiveError> {
-  sys.fs_create_dir_all(to).map_err(|source| {
-    CopyDirRecursiveError::Creating {
-      path: to.to_path_buf(),
-      source,
-    }
-  })?;
-  let read_dir =
-    sys
-      .fs_read_dir(from)
-      .map_err(|source| CopyDirRecursiveError::Reading {
-        path: from.to_path_buf(),
-        source,
-      })?;
-
-  for entry in read_dir {
-    let entry = entry?;
-    let file_type = entry.file_type()?;
-    let new_from = from.join(entry.file_name());
-    let new_to = to.join(entry.file_name());
-
-    if file_type.is_dir() {
-      copy_dir_recursive(sys, &new_from, &new_to).map_err(|source| {
-        CopyDirRecursiveError::Dir {
-          from: new_from.to_path_buf(),
-          to: new_to.to_path_buf(),
-          source: Box::new(source),
-        }
-      })?;
-    } else if file_type.is_file() {
-      sys.fs_copy(&new_from, &new_to).map_err(|source| {
-        CopyDirRecursiveError::Copying {
-          from: new_from.to_path_buf(),
-          to: new_to.to_path_buf(),
-          source,
-        }
-      })?;
-    }
-  }
-
-  Ok(())
-}
-
-pub fn symlink_dir<TSys: sys_traits::BaseFsSymlinkDir>(
-  sys: &TSys,
-  oldpath: &Path,
-  newpath: &Path,
-) -> Result<(), Error> {
-  let err_mapper = |err: Error, kind: Option<ErrorKind>| {
-    Error::new(
-      kind.unwrap_or_else(|| err.kind()),
-      format!(
-        "{}, symlink '{}' -> '{}'",
-        err,
-        oldpath.display(),
-        newpath.display()
-      ),
-    )
-  };
-
-  sys.fs_symlink_dir(oldpath, newpath).map_err(|err| {
-    #[cfg(windows)]
-    if let Some(code) = err.raw_os_error() {
-      if code as u32 == winapi::shared::winerror::ERROR_PRIVILEGE_NOT_HELD
-        || code as u32 == winapi::shared::winerror::ERROR_INVALID_FUNCTION
-      {
-        return err_mapper(err, Some(ErrorKind::PermissionDenied));
-      }
-    }
-    err_mapper(err, None)
-  })
-}
-
 /// Gets the total size (in bytes) of a directory.
 pub fn dir_size(path: &Path) -> std::io::Result<u64> {
   let entries = std::fs::read_dir(path)?;
@@ -338,161 +162,6 @@ pub fn dir_size(path: &Path) -> std::io::Result<u64> {
     };
   }
   Ok(total)
-}
-
-struct LaxSingleProcessFsFlagInner {
-  file_path: PathBuf,
-  fs_file: std::fs::File,
-  finished_token: Arc<tokio_util::sync::CancellationToken>,
-}
-
-impl Drop for LaxSingleProcessFsFlagInner {
-  fn drop(&mut self) {
-    // kill the poll thread
-    self.finished_token.cancel();
-    // release the file lock
-    if let Err(err) = fs3::FileExt::unlock(&self.fs_file) {
-      log::debug!(
-        "Failed releasing lock for {}. {:#}",
-        self.file_path.display(),
-        err
-      );
-    }
-  }
-}
-
-/// A file system based flag that will attempt to synchronize multiple
-/// processes so they go one after the other. In scenarios where
-/// synchronization cannot be achieved, it will allow the current process
-/// to proceed.
-///
-/// This should only be used in places where it's ideal for multiple
-/// processes to not update something on the file system at the same time,
-/// but it's not that big of a deal.
-pub struct LaxSingleProcessFsFlag(
-  #[allow(dead_code)] Option<LaxSingleProcessFsFlagInner>,
-);
-
-impl LaxSingleProcessFsFlag {
-  pub async fn lock(file_path: PathBuf, long_wait_message: &str) -> Self {
-    log::debug!("Acquiring file lock at {}", file_path.display());
-    use fs3::FileExt;
-    let last_updated_path = file_path.with_extension("lock.poll");
-    let start_instant = std::time::Instant::now();
-    let open_result = std::fs::OpenOptions::new()
-      .read(true)
-      .write(true)
-      .create(true)
-      .truncate(false)
-      .open(&file_path);
-
-    match open_result {
-      Ok(fs_file) => {
-        let mut pb_update_guard = None;
-        let mut error_count = 0;
-        while error_count < 10 {
-          let lock_result = fs_file.try_lock_exclusive();
-          let poll_file_update_ms = 100;
-          match lock_result {
-            Ok(_) => {
-              log::debug!("Acquired file lock at {}", file_path.display());
-              let _ignore = std::fs::write(&last_updated_path, "");
-              let token = Arc::new(tokio_util::sync::CancellationToken::new());
-
-              // Spawn a blocking task that will continually update a file
-              // signalling the lock is alive. This is a fail safe for when
-              // a file lock is never released. For example, on some operating
-              // systems, if a process does not release the lock (say it's
-              // killed), then the OS may release it at an indeterminate time
-              //
-              // This uses a blocking task because we use a single threaded
-              // runtime and this is time sensitive so we don't want it to update
-              // at the whims of whatever is occurring on the runtime thread.
-              spawn_blocking({
-                let token = token.clone();
-                let last_updated_path = last_updated_path.clone();
-                move || {
-                  let mut i = 0;
-                  while !token.is_cancelled() {
-                    i += 1;
-                    let _ignore =
-                      std::fs::write(&last_updated_path, i.to_string());
-                    std::thread::sleep(Duration::from_millis(
-                      poll_file_update_ms,
-                    ));
-                  }
-                }
-              });
-
-              return Self(Some(LaxSingleProcessFsFlagInner {
-                file_path,
-                fs_file,
-                finished_token: token,
-              }));
-            }
-            Err(_) => {
-              // show a message if it's been a while
-              if pb_update_guard.is_none()
-                && start_instant.elapsed().as_millis() > 1_000
-              {
-                let pb = ProgressBar::new(ProgressBarStyle::TextOnly);
-                let guard = pb.update_with_prompt(
-                  ProgressMessagePrompt::Blocking,
-                  long_wait_message,
-                );
-                pb_update_guard = Some((guard, pb));
-              }
-
-              // sleep for a little bit
-              tokio::time::sleep(Duration::from_millis(20)).await;
-
-              // Poll the last updated path to check if it's stopped updating,
-              // which is an indication that the file lock is claimed, but
-              // was never properly released.
-              match std::fs::metadata(&last_updated_path)
-                .and_then(|p| p.modified())
-              {
-                Ok(last_updated_time) => {
-                  let current_time = std::time::SystemTime::now();
-                  match current_time.duration_since(last_updated_time) {
-                    Ok(duration) => {
-                      if duration.as_millis()
-                        > (poll_file_update_ms * 2) as u128
-                      {
-                        // the other process hasn't updated this file in a long time
-                        // so maybe it was killed and the operating system hasn't
-                        // released the file lock yet
-                        return Self(None);
-                      } else {
-                        error_count = 0; // reset
-                      }
-                    }
-                    Err(_) => {
-                      error_count += 1;
-                    }
-                  }
-                }
-                Err(_) => {
-                  error_count += 1;
-                }
-              }
-            }
-          }
-        }
-
-        drop(pb_update_guard); // explicit for clarity
-        Self(None)
-      }
-      Err(err) => {
-        log::debug!(
-          "Failed to open file lock at {}. {:#}",
-          file_path.display(),
-          err
-        );
-        Self(None) // let the process through
-      }
-    }
-  }
 }
 
 pub fn specifier_from_file_path(
@@ -658,96 +327,6 @@ mod tests {
         .map(|s| s.to_string())
         .collect::<Vec<_>>(),
       expected
-    );
-  }
-
-  #[tokio::test]
-  async fn lax_fs_lock() {
-    let temp_dir = TempDir::new();
-    let lock_path = temp_dir.path().join("file.lock");
-    let signal1 = Arc::new(Notify::new());
-    let signal2 = Arc::new(Notify::new());
-    let signal3 = Arc::new(Notify::new());
-    let signal4 = Arc::new(Notify::new());
-    tokio::spawn({
-      let lock_path = lock_path.clone();
-      let signal1 = signal1.clone();
-      let signal2 = signal2.clone();
-      let signal3 = signal3.clone();
-      let signal4 = signal4.clone();
-      let temp_dir = temp_dir.clone();
-      async move {
-        let flag =
-          LaxSingleProcessFsFlag::lock(lock_path.to_path_buf(), "waiting")
-            .await;
-        signal1.notify_one();
-        signal2.notified().await;
-        tokio::time::sleep(Duration::from_millis(10)).await; // give the other thread time to acquire the lock
-        temp_dir.write("file.txt", "update1");
-        signal3.notify_one();
-        signal4.notified().await;
-        drop(flag);
-      }
-    });
-    let signal5 = Arc::new(Notify::new());
-    tokio::spawn({
-      let temp_dir = temp_dir.clone();
-      let signal5 = signal5.clone();
-      async move {
-        signal1.notified().await;
-        signal2.notify_one();
-        let flag =
-          LaxSingleProcessFsFlag::lock(lock_path.to_path_buf(), "waiting")
-            .await;
-        temp_dir.write("file.txt", "update2");
-        signal5.notify_one();
-        drop(flag);
-      }
-    });
-
-    signal3.notified().await;
-    assert_eq!(temp_dir.read_to_string("file.txt"), "update1");
-    signal4.notify_one();
-    signal5.notified().await;
-    assert_eq!(temp_dir.read_to_string("file.txt"), "update2");
-  }
-
-  #[tokio::test]
-  async fn lax_fs_lock_ordered() {
-    let temp_dir = TempDir::new();
-    let lock_path = temp_dir.path().join("file.lock");
-    let output_path = temp_dir.path().join("output");
-    let expected_order = Arc::new(Mutex::new(Vec::new()));
-    let count = 10;
-    let mut tasks = Vec::with_capacity(count);
-
-    std::fs::write(&output_path, "").unwrap();
-
-    for i in 0..count {
-      let lock_path = lock_path.clone();
-      let output_path = output_path.clone();
-      let expected_order = expected_order.clone();
-      tasks.push(tokio::spawn(async move {
-        let flag =
-          LaxSingleProcessFsFlag::lock(lock_path.to_path_buf(), "waiting")
-            .await;
-        expected_order.lock().push(i.to_string());
-        // be extremely racy
-        let mut output = std::fs::read_to_string(&output_path).unwrap();
-        if !output.is_empty() {
-          output.push('\n');
-        }
-        output.push_str(&i.to_string());
-        std::fs::write(&output_path, output).unwrap();
-        drop(flag);
-      }));
-    }
-
-    futures::future::join_all(tasks).await;
-    let expected_output = expected_order.lock().join("\n");
-    assert_eq!(
-      std::fs::read_to_string(output_path).unwrap(),
-      expected_output
     );
   }
 }
