+++ resolvers/npm_installer/Cargo.toml
@@ -1,0 +1,53 @@
+# Copyright 2018-2025 the Deno authors. MIT license.
+
+[package]
+name = "deno_npm_installer"
+version = "0.0.1"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+readme = "README.md"
+repository.workspace = true
+description = "Installer of npm packages used in Deno"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+bincode.workspace = true
+boxed_error.workspace = true
+capacity_builder.workspace = true
+deno_config.workspace = true
+deno_error.workspace = true
+deno_npm.workspace = true
+deno_npm_cache.workspace = true
+deno_package_json.workspace = true
+deno_path_util.workspace = true
+deno_resolver.workspace = true
+deno_semver.workspace = true
+deno_terminal.workspace = true
+deno_unsync.workspace = true
+fs3.workspace = true
+futures.workspace = true
+log.workspace = true
+parking_lot.workspace = true
+pathdiff.workspace = true
+rustc-hash.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+sys_traits.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+twox-hash.workspace = true
+url.workspace = true
+
+[target.'cfg(windows)'.dependencies]
+junction.workspace = true
+winapi = { workspace = true, features = ["knownfolders", "mswsock", "objbase", "shlobj", "tlhelp32", "winbase", "winerror", "winsock2"] }
+
+[dev-dependencies]
+sys_traits = { workspace = true, features = ["memory", "real", "serde_json"] }
+test_util.workspace = true
