+++ resolvers/npm_installer/README.md
@@ -1,0 +1,6 @@
+# deno_npm_installer
+
+[![crates](https://img.shields.io/crates/v/deno_npm_installer.svg)](https://crates.io/crates/deno_npm_installer)
+[![docs](https://docs.rs/deno_npm_installer/badge.svg)](https://docs.rs/deno_npm_installer)
+
+Installer for npm packages.
