+++ resolvers/npm_installer/bin_entries.rs
