+++ resolvers/npm_installer/global.rs
@@ -5,32 +5,39 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use async_trait::async_trait;
-use deno_core::futures::stream::FuturesUnordered;
-use deno_core::futures::StreamExt;
 use deno_error::JsErrorBox;
-use deno_lib::util::hash::FastInsecureHasher;
 use deno_npm::NpmResolutionPackage;
 use deno_npm::NpmSystemInfo;
+use deno_npm_cache::NpmCache;
+use deno_npm_cache::NpmCacheHttpClient;
+use deno_npm_cache::NpmCacheSys;
+use deno_npm_cache::TarballCache;
 use deno_resolver::npm::managed::NpmResolutionCell;
+use deno_terminal::colors;
+use futures::stream::FuturesUnordered;
+use futures::StreamExt;
 
-use super::common::NpmPackageFsInstaller;
-use super::PackageCaching;
-use crate::args::LifecycleScriptsConfig;
-use crate::colors;
-use crate::npm::CliNpmCache;
-use crate::npm::CliNpmTarballCache;
+use crate::lifecycle_scripts::LifecycleScripts;
+use crate::lifecycle_scripts::LifecycleScriptsStrategy;
+use crate::LifecycleScriptsConfig;
+use crate::NpmPackageFsInstaller;
+use crate::PackageCaching;
 
 /// Resolves packages from the global npm cache.
-pub struct GlobalNpmPackageInstaller {
-  cache: Arc<CliNpmCache>,
-  tarball_cache: Arc<CliNpmTarballCache>,
+pub struct GlobalNpmPackageInstaller<
+  THttpClient: NpmCacheHttpClient,
+  TSys: NpmCacheSys,
+> {
+  cache: Arc<NpmCache<TSys>>,
+  tarball_cache: Arc<TarballCache<THttpClient, TSys>>,
   resolution: Arc<NpmResolutionCell>,
   lifecycle_scripts: LifecycleScriptsConfig,
   system_info: NpmSystemInfo,
 }
 
-impl std::fmt::Debug for GlobalNpmPackageInstaller {
+impl<THttpClient: NpmCacheHttpClient, TSys: NpmCacheSys> std::fmt::Debug
+  for GlobalNpmPackageInstaller<THttpClient, TSys>
+{
   fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
     f.debug_struct("GlobalNpmPackageInstaller")
       .field("cache", &self.cache)
@@ -42,10 +49,12 @@ impl std::fmt::Debug for GlobalNpmPackageInstaller {
   }
 }
 
-impl GlobalNpmPackageInstaller {
+impl<THttpClient: NpmCacheHttpClient, TSys: NpmCacheSys>
+  GlobalNpmPackageInstaller<THttpClient, TSys>
+{
   pub fn new(
-    cache: Arc<CliNpmCache>,
-    tarball_cache: Arc<CliNpmTarballCache>,
+    cache: Arc<NpmCache<TSys>>,
+    tarball_cache: Arc<TarballCache<THttpClient, TSys>>,
     resolution: Arc<NpmResolutionCell>,
     lifecycle_scripts: LifecycleScriptsConfig,
     system_info: NpmSystemInfo,
@@ -58,10 +67,34 @@ impl GlobalNpmPackageInstaller {
       system_info,
     }
   }
+
+  async fn cache_packages(
+    &self,
+    packages: &[NpmResolutionPackage],
+  ) -> Result<(), deno_npm_cache::EnsurePackageError> {
+    let mut futures_unordered = FuturesUnordered::new();
+    for package in packages {
+      if let Some(dist) = &package.dist {
+        futures_unordered.push(async move {
+          self
+            .tarball_cache
+            .ensure_package(&package.id.nv, dist)
+            .await
+        });
+      }
+    }
+    while let Some(result) = futures_unordered.next().await {
+      // surface the first error
+      result?;
+    }
+    Ok(())
+  }
 }
 
-#[async_trait(?Send)]
-impl NpmPackageFsInstaller for GlobalNpmPackageInstaller {
+#[async_trait::async_trait(?Send)]
+impl<THttpClient: NpmCacheHttpClient, TSys: NpmCacheSys> NpmPackageFsInstaller
+  for GlobalNpmPackageInstaller<THttpClient, TSys>
+{
   async fn cache_packages<'a>(
     &self,
     caching: PackageCaching<'a>,
@@ -75,7 +108,8 @@ impl NpmPackageFsInstaller for GlobalNpmPackageInstaller {
         .subset(&reqs)
         .all_system_packages_partitioned(&self.system_info),
     };
-    cache_packages(&package_partitions.packages, &self.tarball_cache)
+    self
+      .cache_packages(&package_partitions.packages)
       .await
       .map_err(JsErrorBox::from_err)?;
 
@@ -87,14 +121,13 @@ impl NpmPackageFsInstaller for GlobalNpmPackageInstaller {
         .map_err(JsErrorBox::from_err)?;
     }
 
-    let mut lifecycle_scripts =
-      super::common::lifecycle_scripts::LifecycleScripts::new(
-        &self.lifecycle_scripts,
-        GlobalLifecycleScripts::new(
-          self.cache.as_ref(),
-          &self.lifecycle_scripts.root_dir,
-        ),
-      );
+    let mut lifecycle_scripts = LifecycleScripts::new(
+      &self.lifecycle_scripts,
+      GlobalLifecycleScripts::new(
+        self.cache.as_ref(),
+        &self.lifecycle_scripts.root_dir,
+      ),
+    );
 
     // For the global cache, we don't run scripts so we just care that there _are_
     // scripts. Kind of hacky, but avoids fetching the "extra" info from the registry.
@@ -119,33 +152,15 @@ impl NpmPackageFsInstaller for GlobalNpmPackageInstaller {
   }
 }
 
-async fn cache_packages(
-  packages: &[NpmResolutionPackage],
-  tarball_cache: &Arc<CliNpmTarballCache>,
-) -> Result<(), deno_npm_cache::EnsurePackageError> {
-  let mut futures_unordered = FuturesUnordered::new();
-  for package in packages {
-    if let Some(dist) = &package.dist {
-      futures_unordered.push(async move {
-        tarball_cache.ensure_package(&package.id.nv, dist).await
-      });
-    }
-  }
-  while let Some(result) = futures_unordered.next().await {
-    // surface the first error
-    result?;
-  }
-  Ok(())
-}
-
-struct GlobalLifecycleScripts<'a> {
-  cache: &'a CliNpmCache,
+struct GlobalLifecycleScripts<'a, TSys: NpmCacheSys> {
+  cache: &'a NpmCache<TSys>,
   path_hash: u64,
 }
 
-impl<'a> GlobalLifecycleScripts<'a> {
-  fn new(cache: &'a CliNpmCache, root_dir: &Path) -> Self {
-    let mut hasher = FastInsecureHasher::new_without_deno_version();
+impl<'a, TSys: NpmCacheSys> GlobalLifecycleScripts<'a, TSys> {
+  fn new(cache: &'a NpmCache<TSys>, root_dir: &Path) -> Self {
+    use std::hash::Hasher;
+    let mut hasher = twox_hash::XxHash64::default();
     hasher.write(root_dir.to_string_lossy().as_bytes());
     let path_hash = hasher.finish();
     Self { cache, path_hash }
@@ -159,8 +174,8 @@ impl<'a> GlobalLifecycleScripts<'a> {
   }
 }
 
-impl super::common::lifecycle_scripts::LifecycleScriptsStrategy
-  for GlobalLifecycleScripts<'_>
+impl<TSys: NpmCacheSys> LifecycleScriptsStrategy
+  for GlobalLifecycleScripts<'_, TSys>
 {
   fn can_run_scripts(&self) -> bool {
     false
