+++ resolvers/npm_installer/todo.txt
@@ -1,0 +1,4 @@
+This crate is very much a work in progress.
+
+- Use MaybeArc in some places
+- Use sys_traits
